Connect event api
Slight refactor here as well. The event object is now automatically serialized by any
serializer that inherits the mixin. The event objects representation will also include
all of the related file pks and the endpoint where they can be accessed. I'd like a
way to handle that automatically, but this seems straightforward enough for now.