Try ingesting report even if it fails validation.
This is more future proofing than anything. The only way it will fail to ingest
the report is if the serial_number or timestamp are missing. The additional tag will
mostly only be useful in association with the incomplete tag. We'll know the schema
was bad and that's why extraction failed.