[bugfix] SlackApiError can't be serialized
This was causing a bug where if, for instance, a channel couldn't be found we'd get an unhandled exception
and the task results wouldn't be complete. For now I'm just creating a new exception and storing the error
key in it, but this is probably a good case for more robust error info handling as we had discussed a while
back.