Make deployment migration optional
I'm using the celery start script as celery as anything that potentially gets
queued while migrating will wait for celery to start. Also, supervisord is
sensitive to scripts that terminate and will complain in the logs when the
migration finishes if the script exits. On terraform plan, user will be
prompted to set the migrate on deployment flag.