HDSAutoSchema class
We use a custom renderer for all of our JSON responses. This causes issues when auto generating the schema as the default AutoSchema class does not consider the renderer when introspecting and discoering the response format. Here a define a custom HDSAutoSchema class which overrides the  method of  to manually insert our response format for all  responses. I messed around with both  and  to try and find an already written solution to this and could not. I have tested this with an openapi client generator (https://github.com/openapi-generators/openapi-python-client) and confirmed that it correctly builds the response classes and the data is located in the correct keys.