Not sending handled error message to operators
We want to avoid sending a message to the operator
channels on slack, when there is an error that Dory
Handles. The reason for this is that we'd like to
only notify the operators on slack when there is a
necessary piece of information or a direct action item
for them to take. So, we are avoiding sending
handled error messages with this commit.