Harvester load test framework with KMC.
The idea is to mimmic the typical behavior of a harvester, and ramp up to 100 or 1000 harvesters. The methods are stubs right now,
but since errors happen in a somewhat stochastic fashion, I'm using kinetic monte carlo to determine when the next error occurs. I figured
that deserved a commit on it's own