[cleanup] Limit beatbox logs in cloudwath
This is a minor hack I suppose, but if beatbox is going to be running every minute and potentially executing hundreds of requests,
we probably don't want to log them all. We could potentiall write a more sophisticated logger that is more of a rate limiter, but I think
this will do the job fine for now. Failure codes and response are reported to slack anyway