[bugfix] Retry on 500/502/503/504 errors
Beatbox failed once overnight with a 500 code on an endpoint it hits for every test. Typically, these are recoverable. I am adding a a retry up to 5 times for specified server error codes.
The SQS client also retries up to 5 times, so this is in line with what we already do.