[feature] Download grip reports in separate process
The grip reports can be on the order of 10-200MB and this was causing timeouts with the SQS client.
We don't want to increase this timeout and block the main thread. This changes to a paradigm of
accepting the event information, instead of the report directly, then downloading from S3 in a
celery task. This should be the approach for all large pieces of data.