[refactor] clean and simplify log extraction
Log lines no have ascii escape and control sequences which add unwanted characters to the lines, which dont translate well to the frontend.
We were also using many patterns for the lines when we really only needed one for each of can and log files. This also fixes a bug I found where
I put a real log with an exception + traceback in it and the tests failed.

My main reason for doing this though is that we are considering moving to an external API for saving extracted logs, and this allows the methods
to be used in other contexts.