Use redirect instead of direct download link in related_files
Basically if a user opened a report detail view, it would immediately get all of the signed urls
for files from AWS all at once. That means even if they weren't needed, they'd still be created.
Those urls are also on a timer, so they would expire after the window was left open too long. This
just switches to using a new action endpoint which responds with a redirect to the actual download.

I've tested this locally and it seems like there should not be any issue on the frontend side, but
I did only test it with the requests library, and with local files in the media folder, so maybe
we should find a way to test that.