[partial feature] DynamicKey and DynamicKeys classes
An initial implementation of the schema portion for dynamic keys. a JobSchema
object will be aware if it allows dynamic keys or not, and update the payload
schema accordingly. The next commit will handle including this in the form,
validating the dynamic payload when it arrives, and inserting the parameters
for the dynamic payload at time of scheduling.