[feature] Methods for automatic grouping of exception tracebacks
We want to automatically categorize AFTBaseExceptions so we know what problems we are experiencing
most often. This is a pretty brute force approach, but essentially I am stripping the traceback down
to words only, vectorizing it with TfidfVectorizer, and performing hierarchal clustering using
euclidian distance between the vectors. Testing this approach manually seemed to pretty reasonably group
like exceptions