[bugfix/hack] Allow additional properties in scheduler forms
For some reason, the frontend plugin we are using to build/validate the forms is bad at validating the forms.
I can get the payload and validate it manually with the jsonschema library, but it fails to validate even
though i'm filling out the form IT CREATED. Anyway, removing the restriction of not allowing additional
properties seems to make it at least work so we can do cron scheduling.