[bugfix] correctly set the job payload
First there's the bug kosam discovered where we weren't actually using the dynamic payload correctly. It turned out
that it actually was updating the scheduled job in place so it DID update it, but then re-saved it without the
dynamic keys in the job_def. So we need to actually use the payload explicitly after calling .copy()