[Refactor] Install newman locally at project level
To run newman at local level, we need to track package.json as it is
the entry point to run npm commands within the project.
Package-lock.json locks the version installed just in case an update
is available as a patch for newman and so git should not complain
of untracked or uncommitted changes.